% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{dft.gradients}
\alias{dft.gradients}
\title{List of start and end colours for DfT gradient palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
dft.gradients
}
\description{
Gradient colour palettes
}
\details{
Individual palettes returning 10 hex codes for a single shade gradient
}
\keyword{datasets}
