% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{dft.palettes}
\alias{dft.palettes}
\title{Named hex colour palettes using DfT colours}
\format{
An object of class \code{list} of length 6.
}
\usage{
dft.palettes
}
\description{
Colour palettes
}
\keyword{datasets}
