% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayPalette.R
\name{display_palette}
\alias{display_palette}
\title{Launch DfT palette picker tool}
\usage{
display_palette(palette, gradient = FALSE, n = 5)
}
\description{
Displays a visual colour chart of a selected palette
}
