% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayPalette.R
\name{displayPalette}
\alias{displayPalette}
\title{Launch DfT palette picker tool}
\usage{
displayPalette(palette, gradient = FALSE)
}
\description{
Displays a visual colour chart of a selected palette
}
