% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractColours.R
\name{extract_colours}
\alias{extract_colours}
\title{Extract named palette colours as hex codes}
\usage{
extract_colours(...)
}
\arguments{
\item{...}{Individual character names of colours in palette}
}
\description{
Function to extract palette colours as hex codes
}
