% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGradient.R
\name{extract_gradient}
\alias{extract_gradient}
\title{Extract gradient of n colours as hex codes based on DfT palettes}
\usage{
extract_gradient(palette, n = 5)
}
\arguments{
\item{palette}{Gradient palette name as a string, provides the start and end colours of the gradient.}

\item{n}{Number of colours required in created gradient, as an integer. Defaults to 5.}
}
\description{
Function to extract a gradient of colours as hex codes
}
