% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractColours.R
\name{extractColours}
\alias{extractColours}
\title{Extract named palette colours as hex codes}
\usage{
extractColours(...)
}
\arguments{
\item{...}{Individual character names of colours in palette}
}
\description{
Function to extract palette colours as hex codes
}
