% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractColours.R
\name{extractGradient}
\alias{extractGradient}
\title{Extract gradient of n colours as hex codes based on DfT palettes}
\usage{
extractGradient(colours, n = 10)
}
\arguments{
\item{palette}{Palette name as a string}
}
\description{
Function to extract a gradient of colours as hex codes
}
