% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPalettes.R
\name{palette_picker_tool}
\alias{palette_picker_tool}
\title{Launch DfT palette picker tool}
\usage{
palette_picker_tool()
}
\description{
Launches a Shiny tool in R which allows visualisation of the DfT corporate palette in charting functions, to allow creation of accessible charting palettes.
}
