% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleXDft.R
\name{scale_all_dft}
\alias{scale_all_dft}
\title{Extract named palette colours as hex codes}
\usage{
scale_all_dft(palette, gradient, ...)
}
\arguments{
\item{palette}{Named DfT palette to use in the fill.}

\item{gradient}{Boolean response whether to use gradient palette instead of standard palettes.}

\item{...}{Other arguments to pass to extract gradients function}
}
\description{
Internal function to select which palette to use in scale_fill_dft and scale_colour_dft and output warnings/errors as appropriate
}
