% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleXDft.R
\name{scale_fill_dft}
\alias{scale_fill_dft}
\title{Extract named palette colours as hex codes for use in area charts}
\usage{
scale_fill_dft(palette = "main.palette", gradient = FALSE, ...)
}
\arguments{
\item{palette}{Named DfT palette to use in the fill. Defaults to "main.palette"}

\item{gradient}{Boolean response whether to use gradient palette instead of standard palettes. Defaults to false.}

\item{...}{Other arguments to pass to extract gradients function}
}
\description{
Function to use custom DfT palettes for fill colours in ggplot charts
}
