% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeXDft.R
\name{theme_dft}
\alias{theme_dft}
\alias{theme_general_dft}
\title{Displays a visual colour chart of a selected palette}
\usage{
theme_general_dft(
  legend_right = FALSE,
  base_family = "Arial",
  base_size = 14,
  base_line_size = base_size/170
)
}
\arguments{
\item{legend_right}{Logical indicating whether legend should be placed to the right of the plot. If FALSE, the default, legend is positioned below the plot.}

\item{base_family}{Font family}

\item{base_size}{The base font size}

\item{base_line_size}{Default}
}
\description{
Applies a standardised DfT theme to a ggplot.
}
\details{
This theme has been designed using charting best practice principles.
}
