% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeXDft.R
\name{theme_general_dft}
\alias{theme_general_dft}
\title{Displays a visual colour chart of a selected palette}
\usage{
theme_general_dft(
  legend_position = "bottom",
  base_family = "Arial",
  base_size = 14,
  base_line_size = 2,
  margin_sizes = c(2, 1, 1, 1)
)
}
\arguments{
\item{legend_position}{the desired legend position. Selects bottom by default.}

\item{base_family}{Font family}

\item{base_size}{The base font size}

\item{base_line_size}{The base line size}

\item{margin_sizes}{The margin size in cm. Takes a list of the four margins in the order top, right, bottom, left.}
}
\description{
Applies a standardised DfT theme to a ggplot.
}
\details{
This theme has been designed using charting best practice principles.
}
\examples{
## Simple line plot using general theme with default options

library(ggplot2)

df <- economics_long[economics_long$variable \%in\% c("psavert", "uempmed"), ]

# plot
ggplot(df, aes(x=date)) +
  geom_line(aes(y=value, col=variable), size = 1) +
  theme_general_dft()
}
