% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeXDft.R
\name{theme_line_dft}
\alias{theme_line_dft}
\title{Applies a standardised DfT theme to a ggplot line plot}
\usage{
theme_line_dft(
  legend_position = "bottom",
  base_family = "",
  base_size = 14,
  base_line_size = 2,
  accuracy = 1,
  palette = "main.palette",
  gradient = FALSE,
  ...
)
}
\arguments{
\item{legend_position}{the desired legend position. Selects bottom by default.}

\item{base_family}{Font family}

\item{base_size}{The base font size}

\item{base_line_size}{The base line size}

\item{accuracy}{The accuracy of the y-axis, default to 1 for 0 decimal place. 0.1 would give 1 decimal place, 0.01 would give 2 decimal places.}

\item{palette}{The name of the DfT palette you want to use}

\item{gradient}{Boolean response whether to use gradient palette instead of standard palettes. Defaults to false.}

\item{...}{Other arguments to pass to extract gradients function}
}
\description{
Applies a standardised DfT theme to a ggplot line plot.
}
\details{
This theme has been designed using charting best practice principles.
}
\examples{
## Simple line plot using standard line theme with default options

library(ggplot2)

df <- economics_long[economics_long$variable \%in\% c("psavert", "uempmed"), ]

# plot
ggplot(df, aes(x=date, y=value, col=variable, label=variable)) +
  geom_line(size = 1) +
  theme_line_dft()
## More complex line plot selecting another dft palette

library(ggplot2)

df <- economics_long[economics_long$variable \%in\% c("psavert", "uempmed"), ]

# plot
ggplot(df, aes(x=date, y=value, col=variable, label=variable)) +
  geom_line(size = 1) +
  theme_line_dft(palette = "clear.skies")
## More complex line plot using a gradient palette with two groups

library(ggplot2)

df <- economics_long[economics_long$variable \%in\% c("psavert", "uempmed"), ]

# plot
ggplot(df, aes(x=date, y=value, col=variable, label=variable)) +
  geom_line(size = 1) +
  theme_line_dft(gradient = TRUE, n=2)
}
