% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayPalette.R
\name{display_palette}
\alias{display_palette}
\title{Displays a visual colour chart of a selected palette}
\usage{
display_palette(palette, gradient = FALSE, ...)
}
\arguments{
\item{palette}{string, name of palette of interest}

\item{gradient}{boolean to indicate whether palette should be standard colours or gradient. Default is gradient = FALSE}

\item{...}{Additional arguments to pass to the extract_gradient function, such as number of colours to display in gradient palette}
}
\description{
Displays a visual colour chart of a selected palette based on the name of the palette.
}
\details{
Current DfT palettes available are main.palette, electric.brights, just.beachy, cycling.hills, mountain.train and clear.skies.

All palettes are available as either pre-set palettes of colours, or as gradients. Gradient palettes allow the user to select as many colour options as they like using the dots argument.
}
