% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGradient.R
\name{extractGradient}
\alias{extractGradient}
\title{Extract gradient of n colours as hex codes based on DfT palettes}
\usage{
extractGradient(colours, n = 5)
}
\arguments{
\item{colours}{Individual names of colours to start and end gradient with, in a vector}

\item{n}{Number of colours required in created gradient, as an integer. Defaults to 5.}
}
\description{
Function to extract a gradient of colours as hex codes
}
