% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleXDft.R
\name{scale_colour_dft}
\alias{scale_colour_dft}
\title{Extract named palette colours as hex codes for use in line charts}
\usage{
scale_colour_dft(palette = "electric.brights", gradient = FALSE, ...)
}
\arguments{
\item{palette}{Named DfT palette to use in line colours. Defaults to "electric.brights"}

\item{gradient}{Boolean response whether to use gradient palette instead of standard palettes. Defaults to false.}

\item{...}{Other arguments to pass to extract gradients function}
}
\description{
Function to use custom DfT palettes for line colours in ggplot charts
}
\examples{
## Simple line plot using default palette

library(ggplot2)

df <- economics_long[economics_long$variable \%in\% c("psavert", "uempmed"), ]

# plot
ggplot(df, aes(x=date)) +
  geom_line(aes(y=value, col=variable), size = 1) +
  scale_colour_dft()+
  theme(axis.text.x = element_text(angle = 90, vjust=0.5, size = 8))  # turn off minor grid
}
