% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeXDft.R
\name{theme_line_dft}
\alias{theme_line_dft}
\title{Applies a standardised DfT theme to a ggplot line plot}
\usage{
theme_line_dft(
  legend_right = FALSE,
  base_family = "",
  base_size = 14,
  base_line_size = base_size/170,
  accuracy = 1,
  palette = "main.palette",
  gradient = FALSE,
  ...
)
}
\arguments{
\item{legend_right}{Logical indicating whether legend should be placed to the right of the plot. If FALSE, the default, legend is positioned below the plot.}

\item{base_family}{Font family}

\item{base_size}{The base font size}

\item{base_line_size}{Default}

\item{accuracy}{The accuracy of the y-axis, default to 1 for 0 decimal place. 0.1 would give 1 decimal place.}

\item{palette}{The name of the DfT palette you want to use}

\item{gradient}{Boolean response whether to use gradient palette instead of standard palettes. Defaults to false.}

\item{...}{Other arguments to pass to extract gradients function}
}
\description{
Applies a standardised DfT theme to a ggplot line plot.
}
\details{
This theme has been designed using charting best practice principles.
}
